requirements.txt
imageio==2.24.0
ipython==8.8.0
kornia==0.6.9
matplotlib==3.6.2
moviepy==1.0.3
numpy==1.24.1
omegaconf==2.3.0
Pillow==9.4.0
pydub==0.25.1
pyiqa==0.1.5
pytorch_lightning==1.6.5
streamlit==1.16.0
timm==0.4.5
torch==1.12.0+cu113 
torchaudio==0.12.0+cu113
torchmetrics==0.10.2
torchvision==0.13.0+cu113
tqdm==4.64.1
transformers==4.25.1
wget==3.2
imageio-ffmpeg==0.4.7
ftfy==6.1.1
moviepy==1.0.3
decorator==4.4.2
ipywidgets==8.0.2
gTTS==2.2.4
git+https://github.com/openai/CLIP.git
